+++ extended_mypy_django_plugin/plugin/_known_annotations.py
@@ -7,5 +7,4 @@ class KnownClasses(enum.Enum):
 
 class KnownAnnotations(enum.Enum):
     CONCRETE = "extended_mypy_django_plugin.annotations.Concrete"
-    CONCRETE_QUERYSET = "extended_mypy_django_plugin.annotations.ConcreteQuerySet"
     DEFAULT_QUERYSET = "extended_mypy_django_plugin.annotations.DefaultQuerySet"