+++ docs/api/changelog.rst
@@ -12,6 +12,7 @@ Changelog
     * The concrete annotations understand ``type[Annotation[inner]]`` and ``Annotation[type[inner]]``
       better now and will do the right thing
     * When an annotation would transform into a Union of one item, now it becomes that one item
+    * Removed ``ConcreteQuerySet`` and made ``DefaultQuerySet`` take on that functionality
 
 .. _release-0.5.3: